mandalinka/requirements.txt
asgiref==3.5.2
autopep8==1.7.0
click==8.1.3
crispy-bootstrap5==0.6
debugpy>=1.6.3
Django==4.1
django-crispy-forms==1.14.0
django-node-assets==0.9.11
joblib==1.1.0
nltk==3.7
Pillow==9.2.0
psycopg2-binary==2.9.5
pycodestyle==2.9.1
regex==2022.8.17
ruby==0.0.2
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tqdm==4.64.1
